@@ -1,26 +1,78 @@
 [tool.poetry]
 authors = ["widal001 <billy.daly@agile6.com>"]
-description = "Python package for analyzing data related to the Grants Equity Project"
+description = "Python package for analyzing data related to the Simpler Grants Project"
 name = "analytics"
 packages = [{include = "analytics", from = "src"}]
 readme = "README.md"
 version = "0.1.0"
 
+[tool.poetry.scripts]
+analytics = "analytics.cli:app"
+
 [tool.poetry.dependencies]
+dynaconf = "^3.2.4"
+kaleido = "0.2.1"
 notebook = "^7.0.0" # Goal is to replace this with another method of presenting charts
-pandas = "^2.0.3" 
-plotly = "^5.15.0" 
-pydantic = "^2.0.3" 
-python = "^3.10" 
+pandas = "^2.0.3"
 pandas-stubs = "^2.0.2.230605"
+plotly = "^5.15.0"
+pydantic = "^2.0.3"
+python = "^3.11"
+slack-sdk = "^3.23.0"
+typer = {extras = ["all"], version = "^0.9.0"}
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.7.0"
 mypy = "^1.4.1"
+pylint = "^3.0.2"
 pytest = "^7.4.0"
+pytest-cov = "^4.1.0"
 ruff = "^0.0.278"
 safety = "^2.3.5"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
 requires = ["poetry-core"]
+
+[tool.mypy]
+python_version = "3.11"
+
+[[tool.mypy.overrides]]
+ignore_missing_imports = true
+module = [
+  "plotly.*",
+  "dynaconf.*",
+]
+
+[tool.pylint."MESSAGE CONTROL"]
+disable = [
+  "R0801", # duplicate-code
+  "W0511", # fix-me
+  "R0913", # too-many-arguments
+  "R0902", # too-many-instance-attributes
+  "R0903", # too-few-public-methods
+]
+
+[tool.ruff]
+ignore = [
+  "ANN101", # missing type annotation for self
+  "ANN102", # missing type annotation for cls
+  "D203", # no blank line before class
+  "D212", # multi-line summary first line
+  "FIX002", # line contains TODO
+  "PD901", # pandas df variable name
+  "PLR0913", # Too many arguments to function call
+  "PTH123", # `open()` should be replaced by `Path.open()`
+  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
+  "TD003", # missing an issue link on TODO
+  "T201", # use of `print` detected
+]
+line-length = 100
+select = ["ALL"]
+
+[tool.pytest.ini_options]
+filterwarnings = [
+  # kaleido is throwing a Deprecation warning in one of its dependencies
+  # TODO(widal001): 2022-12-12 - Try removing after Kaleido issues a new release
+  'ignore:.*setDaemon\(\) is deprecated.*:DeprecationWarning',
+]
