@@ -1,6 +1,7 @@
+"""Dummy test to make sure the package is installed correctly."""
 import analytics
 
 
 def test_package_import_and_name():
-    """The package should be imported and named correctly"""
+    """The package should be imported and named correctly."""
     assert analytics.__name__ == "analytics"
