@@ -9,16 +9,16 @@ logger = logging.getLogger(__name__)
 
 
 class PostgresDBConfig(PydanticBaseEnvConfig):
-    check_connection_on_init: bool = Field(True, env="DB_CHECK_CONNECTION_ON_INIT")
-    aws_region: Optional[str] = Field(None, env="AWS_REGION")
-    host: str = Field("localhost", env="DB_HOST")
-    name: str = Field("main-db", env="DB_NAME")
-    username: str = Field("local_db_user", env="DB_USER")
-    password: Optional[str] = Field(None, env="DB_PASSWORD")
-    db_schema: str = Field("public", env="DB_SCHEMA")
-    port: int = Field(5432, env="DB_PORT")
-    hide_sql_parameter_logs: bool = Field(True, env="HIDE_SQL_PARAMETER_LOGS")
-    ssl_mode: str = Field("require", env="DB_SSL_MODE")
+    check_connection_on_init: bool = Field(True, alias="DB_CHECK_CONNECTION_ON_INIT")
+    aws_region: Optional[str] = Field(None, alias="AWS_REGION")
+    host: str = Field(alias="DB_HOST")
+    name: str = Field(alias="DB_NAME")
+    username: str = Field(alias="DB_USER")
+    password: Optional[str] = Field(None, alias="DB_PASSWORD")
+    db_schema: str = Field("public", alias="DB_SCHEMA")
+    port: int = Field(5432, alias="DB_PORT")
+    hide_sql_parameter_logs: bool = Field(True, alias="HIDE_SQL_PARAMETER_LOGS")
+    ssl_mode: str = Field("require", alias="DB_SSL_MODE")
 
 
 def get_db_config() -> PostgresDBConfig:
