@@ -1,19 +1,27 @@
-from apiflask import fields
+from src.api.schemas.extension import Schema, fields
+from src.pagination.pagination_schema import PaginationInfoSchema
 
-from src.api.schemas import request_schema
 
-
-class ValidationErrorSchema(request_schema.OrderedSchema):
+class ValidationIssueSchema(Schema):
     type = fields.String(metadata={"description": "The type of error"})
     message = fields.String(metadata={"description": "The message to return"})
-    rule = fields.String(metadata={"description": "The rule that failed"})
     field = fields.String(metadata={"description": "The field that failed"})
-    value = fields.String(metadata={"description": "The value that failed"})
 
 
-class ResponseSchema(request_schema.OrderedSchema):
+class BaseResponseSchema(Schema):
     message = fields.String(metadata={"description": "The message to return"})
-    data = fields.Field(metadata={"description": "The REST resource object"}, dump_default={})
+    data = fields.MixinField(metadata={"description": "The REST resource object"}, dump_default={})
     status_code = fields.Integer(metadata={"description": "The HTTP status code"}, dump_default=200)
-    warnings = fields.List(fields.Nested(ValidationErrorSchema), dump_default=[])
-    errors = fields.List(fields.Nested(ValidationErrorSchema), dump_default=[])
+
+
+class ErrorResponseSchema(BaseResponseSchema):
+    errors = fields.List(fields.Nested(ValidationIssueSchema()), dump_default=[])
+
+
+class ResponseSchema(BaseResponseSchema):
+    pagination_info = fields.Nested(
+        PaginationInfoSchema(),
+        metadata={"description": "The pagination information for paginated endpoints"},
+    )
+
+    warnings = fields.List(fields.Nested(ValidationIssueSchema()), dump_default=[])
