@@ -6,6 +6,7 @@ import sqlalchemy
 
 import src.adapters.db as db
 import src.logging
+from src.adapters.db.type_decorators.postgres_type_decorators import StrEnumColumn
 from src.db.models import metadata
 
 # this is the Alembic Config object, which provides
@@ -16,7 +17,6 @@ logger = logging.getLogger("migrations")
 
 # Initialize logging
 with src.logging.init("migrations"):
-
     # add your model's MetaData object here
     # for 'autogenerate' support
     # from myapp import mymodel
@@ -30,7 +30,7 @@ with src.logging.init("migrations"):
 
     def include_object(
         object: sqlalchemy.schema.SchemaItem,
-        name: str,
+        name: str | None,
         type_: str,
         reflected: bool,
         compare_to: Any,
@@ -39,6 +39,16 @@ with src.logging.init("migrations"):
             return False
         else:
             return True
+
+    def render_item(type_: str, obj: Any, autogen_context: Any) -> Any:
+        # Alembic tries to set the type of the column as StrEnumColumn
+        # despite it being derived from the Text column type,
+        # so force it to be Text during it's generation process
+        if type_ == "type" and isinstance(obj, StrEnumColumn):
+            return "sa.Text()"
+
+        # False means to use the default processing
+        return False
 
     def run_migrations_online() -> None:
         """Run migrations in 'online' mode.
@@ -57,6 +67,7 @@ with src.logging.init("migrations"):
                 include_schemas=False,
                 include_object=include_object,
                 compare_type=True,
+                render_item=render_item,
             )
             with context.begin_transaction():
                 context.run_migrations()
