@@ -1,6 +1,6 @@
 import logging
 
-from . import base, user_models
+from . import base, opportunity_models
 
 logger = logging.getLogger(__name__)
 
@@ -8,4 +8,4 @@ logger = logging.getLogger(__name__)
 # This is used by tests to create the test database.
 metadata = base.metadata
 
-__all__ = ["metadata", "user_models"]
+__all__ = ["metadata", "opportunity_models"]
