@@ -101,7 +101,7 @@ class HumanReadableFormatter(logging.Formatter):
     def format(self, record: logging.LogRecord) -> str:
         message = super().format(record)
         return decodelog.format_line(
-            datetime.utcfromtimestamp(record.created),
+            datetime.fromtimestamp(record.created),
             record.name,
             record.funcName,
             record.levelname,
