@@ -3,6 +3,8 @@ import contextlib
 import logging
 import uuid
 
+from sqlalchemy import text
+
 import src.adapters.db as db
 from src.adapters.db.clients.postgres_config import get_db_config
 
@@ -18,10 +20,6 @@ def create_isolated_db(monkeypatch) -> db.DBClient:
     """
     schema_name = f"test_schema_{uuid.uuid4().int}"
     monkeypatch.setenv("DB_SCHEMA", schema_name)
-    monkeypatch.setenv("DB_NAME", "postgres")
-    monkeypatch.setenv("DB_USER", "postgres")
-    monkeypatch.setenv("DB_PASSWORD", "secret123")
-    monkeypatch.setenv("ENVIRONMENT", "local")
     monkeypatch.setenv("DB_CHECK_CONNECTION_ON_INIT", "False")
 
     # To improve test performance, don't check the database connection
@@ -29,8 +27,10 @@ def create_isolated_db(monkeypatch) -> db.DBClient:
     db_client = db.PostgresDBClient()
     with db_client.get_connection() as conn:
         _create_schema(conn, schema_name)
+
         try:
             yield db_client
+
         finally:
             _drop_schema(conn, schema_name)
 
@@ -39,11 +39,15 @@ def _create_schema(conn: db.Connection, schema_name: str):
     """Create a database schema."""
     db_test_user = get_db_config().username
 
-    conn.execute(f"CREATE SCHEMA IF NOT EXISTS {schema_name} AUTHORIZATION {db_test_user};")
+    with conn.begin():
+        conn.execute(
+            text(f"CREATE SCHEMA IF NOT EXISTS {schema_name} AUTHORIZATION {db_test_user};")
+        )
     logger.info("create schema %s", schema_name)
 
 
 def _drop_schema(conn: db.Connection, schema_name: str):
     """Drop a database schema."""
-    conn.execute(f"DROP SCHEMA {schema_name} CASCADE;")
+    with conn.begin():
+        conn.execute(text(f"DROP SCHEMA {schema_name} CASCADE;"))
     logger.info("drop schema %s", schema_name)
