@@ -1,4 +1,5 @@
 import logging
+from dataclasses import dataclass
 
 import pytest
 
@@ -6,32 +7,30 @@ from src.adapters.db.clients.postgres_client import get_connection_parameters, v
 from src.adapters.db.clients.postgres_config import get_db_config
 
 
-class DummyConnectionInfo:
-    def __init__(self, ssl_in_use, attributes):
-        self.ssl_in_use = ssl_in_use
-        self.attributes = attributes
-        self.ssl_attribute_names = tuple(attributes.keys())
+@dataclass
+class DummyPgConn:
+    ssl_in_use: bool
 
-    def ssl_attribute(self, name):
-        return self.attributes[name]
+
+class DummyConnectionInfo:
+    def __init__(self, ssl_in_use):
+        self.pgconn = DummyPgConn(ssl_in_use)
 
 
 def test_verify_ssl(caplog):
     caplog.set_level(logging.INFO)  # noqa: B1
 
-    conn_info = DummyConnectionInfo(True, {"protocol": "ABCv3", "key_bits": "64", "cipher": "XYZ"})
+    conn_info = DummyConnectionInfo(True)
     verify_ssl(conn_info)
 
-    assert caplog.messages == [
-        "database connection is using SSL: protocol ABCv3, key_bits 64, cipher XYZ"
-    ]
+    assert caplog.messages == ["database connection is using SSL"]
     assert caplog.records[0].levelname == "INFO"
 
 
 def test_verify_ssl_not_in_use(caplog):
     caplog.set_level(logging.INFO)  # noqa: B1
 
-    conn_info = DummyConnectionInfo(False, {})
+    conn_info = DummyConnectionInfo(False)
     verify_ssl(conn_info)
 
     assert caplog.messages == ["database connection is not using SSL"]
