@@ -1,23 +1,31 @@
 export const ExternalRoutes = {
   ABOUT_HHS: "https://www.hhs.gov/about/index.html",
-  ACCESSIBILITY_COMPLIANCE:
-    "https://www.grants.gov/web/grants/accessibility-compliance.html",
+  ACCESSIBILITY_COMPLIANCE: "https://www.grants.gov/accessibility-compliance",
+  EMAIL_SIMPLERGRANTSGOV: "simpler@grants.gov",
+  EMAIL_SUPPORT: "support@grants.gov",
   FOIA: "https://www.hhs.gov/foia/index.html",
-  NO_FEAR: "https://www.hhs.gov/about/agencies/asa/eeo/no-fear-act/index.html",
-  INSPECTOR_GENERAL: "https://oig.hhs.gov/",
-  PERFORMANCE_REPORTS: "https://www.hhs.gov/about/budget/index.html",
-  PRIVACY_POLICY: "https://www.grants.gov/web/grants/privacy.html",
-  USA: "https://www.usa.gov",
-  HHS: "https://www.hhs.gov",
-  MILESTONES:
-    "https://github.com/HHS/grants-equity/blob/main/documentation/milestones/milestone_short_descriptions.md",
-  GITHUB_REPO: "https://github.com/HHS/grants-equity",
-  CONTACT_EMAIL: "mailto:support@grants.gov",
+  GITHUB_REPO: "https://github.com/HHS/simpler-grants-gov",
+  GITHUB_REPO_CONTRIBUTING:
+    "https://github.com/HHS/simpler-grants-gov/blob/main/CONTRIBUTING.md",
+  GITHUB_REPO_GOALS:
+    "https://github.com/HHS/simpler-grants-gov/blob/main/documentation/goals.md",
+  GITHUB_REPO_ROADMAP: "https://github.com/orgs/HHS/projects/12/views/2",
+  GITHUB_ADR:
+    "https://github.com/HHS/simpler-grants-gov/tree/main/documentation/decisions",
   GRANTS_HOME: "https://www.grants.gov",
-  GRANTS_TWITTER: "https://twitter.com/grantsdotgov",
-  GRANTS_YOUTUBE: "https://www.youtube.com/user/GrantsGovUS",
-  GRANTS_RSS: "https://www.grants.gov/web/grants/rss.html",
   GRANTS_NEWSLETTER:
     "https://www.grants.gov/web/grants/connect/newsletter-archive.html",
+  GRANTS_RSS: "https://grants.gov/connect/rss-feeds",
+  GRANTS_TWITTER: "https://twitter.com/grantsdotgov",
+  GRANTS_YOUTUBE: "https://www.youtube.com/user/GrantsGovUS",
   GRANTS_BLOG: "https://grantsgovprod.wordpress.com/",
+  HHS: "https://www.hhs.gov",
+  INSPECTOR_GENERAL: "https://oig.hhs.gov/",
+  MILESTONE_SEARCH_MVP: "https://github.com/HHS/simpler-grants-gov/issues/89",
+  MILESTONE_GET_OPPORTUNITIES:
+    "https://github.com/HHS/simpler-grants-gov/issues/70",
+  NO_FEAR: "https://www.hhs.gov/about/agencies/asa/eeo/no-fear-act/index.html",
+  PERFORMANCE_REPORTS: "https://www.hhs.gov/about/budget/index.html",
+  PRIVACY_POLICY: "https://www.grants.gov/privacy-policy",
+  USA: "https://www.usa.gov",
 };
