@@ -1,4 +1,4 @@
-import type { GetServerSideProps, NextPage } from "next";
+import type { GetStaticProps, NextPage } from "next";
 
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import React from "react";
@@ -7,7 +7,7 @@ const Health: NextPage = () => {
   return <>healthy</>;
 };
 
-export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
+export const getStaticProps: GetStaticProps = async ({ locale }) => {
   const translations = await serverSideTranslations(locale ?? "en");
   return { props: { ...translations } };
 };
