@@ -1,6 +1,5 @@
 import { Meta } from "@storybook/react";
-
-import GoalContent from "src/components/GoalContent";
+import GoalContent from "src/pages/content/IndexGoalContent";
 
 const meta: Meta<typeof GoalContent> = {
   title: "Components/Content/Goal Content",
