@@ -1,6 +1,5 @@
 import { render, screen } from "@testing-library/react";
-
-import WtGIContent from "src/components/WtGIContent";
+import WtGIContent from "src/pages/content/WtGIContent";
 
 describe("Ways to get involved", () => {
   it("Renders without errors", () => {
