@@ -13,13 +13,13 @@ import (
 	"github.com/gruntwork-io/terratest/modules/terraform"
 )
 
+var uniqueId = strings.ToLower(random.UniqueId())
+var workspaceName = fmt.Sprintf("t-%s", uniqueId)
 var appName = flag.String("app_name", "", "name of subdirectory that holds the app's infrastructure code")
 
-func TestDev(t *testing.T) {
+func TestService(t *testing.T) {
 	BuildAndPublish(t)
 
-	uniqueId := strings.ToLower(random.UniqueId())
-	workspaceName := fmt.Sprintf("t-%s", uniqueId)
 	imageTag := shell.RunCommandAndGetOutput(t, shell.Command{
 		Command:    "git",
 		Args:       []string{"rev-parse", "HEAD"},
@@ -28,61 +28,65 @@ func TestDev(t *testing.T) {
 	terraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
 		Reconfigure:  true,
 		TerraformDir: fmt.Sprintf("../%s/service/", *appName),
-		VarFiles:     []string{"dev.tfvars"},
 		Vars: map[string]interface{}{
-			"image_tag": imageTag,
+			"environment_name": "dev",
+			"image_tag":        imageTag,
 		},
 	})
 
-	defer DestroyDevEnvironmentAndWorkspace(t, terraformOptions, workspaceName)
-	CreateDevEnvironmentInWorkspace(t, terraformOptions, workspaceName)
+	fmt.Println("::group::Initialize service module")
+	TerraformInit(t, terraformOptions, "dev.s3.tfbackend")
+	fmt.Println("::endgroup::")
+
+	defer terraform.WorkspaceDelete(t, terraformOptions, workspaceName)
+	fmt.Println("::group::Select new terraform workspace")
+	terraform.WorkspaceSelectOrNew(t, terraformOptions, workspaceName)
+	fmt.Println("::endgroup::")
+
+	defer DestroyService(t, terraformOptions)
+	fmt.Println("::group::Create service layer")
+	terraform.Apply(t, terraformOptions)
+	fmt.Println("::endgroup::")
+
 	WaitForServiceToBeStable(t, workspaceName)
 	RunEndToEndTests(t, terraformOptions)
 }
 
 func BuildAndPublish(t *testing.T) {
+	fmt.Println("::group::Initialize build-repository module")
 	// terratest currently does not support passing a file as the -backend-config option
 	// so we need to manually call terraform rather than using terraform.Init
 	// see https://github.com/gruntwork-io/terratest/issues/517
 	// it looks like this PR would add functionality for this: https://github.com/gruntwork-io/terratest/pull/558
 	// after which we add BackendConfig: []string{"dev.s3.tfbackend": terraform.KeyOnly} to terraformOptions
 	// and replace the call to terraform.RunTerraformCommand with terraform.Init
-	terraform.RunTerraformCommand(t, &terraform.Options{
+	TerraformInit(t, &terraform.Options{
 		TerraformDir: fmt.Sprintf("../%s/build-repository/", *appName),
-	}, "init", "-backend-config=shared.s3.tfbackend")
+	}, "shared.s3.tfbackend")
+	fmt.Println("::endgroup::")
 
+	fmt.Println("::group::Build release")
 	shell.RunCommand(t, shell.Command{
 		Command:    "make",
 		Args:       []string{"release-build", fmt.Sprintf("APP_NAME=%s", *appName)},
 		WorkingDir: "../../",
 	})
+	fmt.Println("::endgroup::")
 
+	fmt.Println("::group::Publish release")
 	shell.RunCommand(t, shell.Command{
 		Command:    "make",
 		Args:       []string{"release-publish", fmt.Sprintf("APP_NAME=%s", *appName)},
 		WorkingDir: "../../",
 	})
-}
-
-func CreateDevEnvironmentInWorkspace(t *testing.T, terraformOptions *terraform.Options, workspaceName string) {
-	fmt.Printf("::group::Create dev environment in new workspace '%s\n'", workspaceName)
-
-	// terratest currently does not support passing a file as the -backend-config option
-	// so we need to manually call terraform rather than using terraform.Init
-	// see https://github.com/gruntwork-io/terratest/issues/517
-	// it looks like this PR would add functionality for this: https://github.com/gruntwork-io/terratest/pull/558
-	// after which we add BackendConfig: []string{"dev.s3.tfbackend": terraform.KeyOnly} to terraformOptions
-	// and replace the call to terraform.RunTerraformCommand with terraform.Init
-	terraform.RunTerraformCommand(t, terraformOptions, "init", "-backend-config=dev.s3.tfbackend")
-	terraform.WorkspaceSelectOrNew(t, terraformOptions, workspaceName)
-	terraform.Apply(t, terraformOptions)
 	fmt.Println("::endgroup::")
 }
 
 func WaitForServiceToBeStable(t *testing.T, workspaceName string) {
 	fmt.Println("::group::Wait for service to be stable")
+	appName := *appName
 	environmentName := "dev"
-	serviceName := fmt.Sprintf("%s-%s-%s", workspaceName, *appName, environmentName)
+	serviceName := fmt.Sprintf("%s-%s-%s", workspaceName, appName, environmentName)
 	shell.RunCommand(t, shell.Command{
 		Command:    "aws",
 		Args:       []string{"ecs", "wait", "services-stable", "--cluster", serviceName, "--services", serviceName},
@@ -100,14 +104,14 @@ func RunEndToEndTests(t *testing.T, terraformOptions *terraform.Options) {
 	fmt.Println("::endgroup::")
 }
 
-func EnableDestroy(t *testing.T, terraformOptions *terraform.Options, workspaceName string) {
-	fmt.Println("::group::Setting force_destroy = true and prevent_destroy = false for s3 buckets")
+func EnableDestroyService(t *testing.T, terraformOptions *terraform.Options) {
+	fmt.Println("::group::Set force_destroy = true and prevent_destroy = false for s3 buckets in service layer")
 	shell.RunCommand(t, shell.Command{
 		Command: "sed",
 		Args: []string{
 			"-i.bak",
 			"s/force_destroy = false/force_destroy = true/g",
-			"infra/modules/service/access_logs.tf",
+			"infra/modules/service/access-logs.tf",
 		},
 		WorkingDir: "../../",
 	})
@@ -116,19 +120,17 @@ func EnableDestroy(t *testing.T, terraformOptions *terraform.Options, workspaceN
 		Args: []string{
 			"-i.bak",
 			"s/prevent_destroy = true/prevent_destroy = false/g",
-			"infra/modules/service/access_logs.tf",
+			"infra/modules/service/access-logs.tf",
 		},
 		WorkingDir: "../../",
 	})
-	terraform.RunTerraformCommand(t, terraformOptions, "init", "-backend-config=dev.s3.tfbackend")
 	terraform.Apply(t, terraformOptions)
+	fmt.Println("::endgroup::")
 }
 
-func DestroyDevEnvironmentAndWorkspace(t *testing.T, terraformOptions *terraform.Options, workspaceName string) {
-	EnableDestroy(t, terraformOptions, workspaceName)
-	fmt.Println("::group::Destroy environment and workspace")
-	terraform.RunTerraformCommand(t, terraformOptions, "init", "-backend-config=dev.s3.tfbackend")
+func DestroyService(t *testing.T, terraformOptions *terraform.Options) {
+	EnableDestroyService(t, terraformOptions)
+	fmt.Println("::group::Destroy service layer")
 	terraform.Destroy(t, terraformOptions)
-	terraform.WorkspaceDelete(t, terraformOptions, workspaceName)
 	fmt.Println("::endgroup::")
 }
